+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project's packages adheres to [Semantic Versioning](http://semver.org/s
 
 ## [Unreleased]
 
+## [2.45.1] - 2023-10-26
+
 ### Changed
 
 - Upgrade K8s dependencies ([#1149](https://github.com/giantswarm/kubectl-gs/pull/1149)).
@@ -1339,7 +1341,8 @@ This release supports rendering for CRs:
 - `AppCatalog`
 - `App`
 
-[Unreleased]: https://github.com/giantswarm/kubectl-gs/compare/v2.45.0...HEAD
+[Unreleased]: https://github.com/giantswarm/kubectl-gs/compare/v2.45.1...HEAD
+[2.45.1]: https://github.com/giantswarm/kubectl-gs/compare/v2.45.0...v2.45.1
 [2.45.0]: https://github.com/giantswarm/kubectl-gs/compare/v2.44.0...v2.45.0
 [2.44.0]: https://github.com/giantswarm/kubectl-gs/compare/v2.43.0...v2.44.0
 [2.43.0]: https://github.com/giantswarm/kubectl-gs/compare/v2.42.0...v2.43.0